Advances in parsing nucleotide changes.
