+++ jannovar-hgvs/src/test/java/de/charite/compbio/jannovar/hgvs/parser/TestProteinParsing.java
@@ -1,48 +1,13 @@
 package de.charite.compbio.jannovar.hgvs.parser;
 
-import org.antlr.v4.runtime.ANTLRInputStream;
-import org.antlr.v4.runtime.BaseErrorListener;
-import org.antlr.v4.runtime.CommonTokenStream;
-import org.antlr.v4.runtime.RecognitionException;
-import org.antlr.v4.runtime.Recognizer;
 import org.junit.Test;
-
-import de.charite.compbio.jannovar.hgvs.parser.HGVSParser.Hgvs_variantContext;
 
 /**
  * Systematically test parsing of protein changes
  *
  * @author Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>
  */
-public class TestProteinParsing {
-
-	protected Hgvs_variantContext parseString(String inputString) {
-		return parseString(inputString, false);
-	}
-
-	protected Hgvs_variantContext parseString(String inputString, boolean trace) {
-		if (trace) {
-			ANTLRInputStream inputStream = new ANTLRInputStream(inputString);
-			HGVSLexer l = new HGVSLexer(inputStream);
-			System.err.println(l.getAllTokens());
-		}
-		ANTLRInputStream inputStream = new ANTLRInputStream(inputString);
-		HGVSLexer l = new HGVSLexer(inputStream);
-		HGVSParser p = new HGVSParser(new CommonTokenStream(l));
-		p.setTrace(trace);
-		p.addErrorListener(new BaseErrorListener() {
-			@Override
-			public void syntaxError(Recognizer<?, ?> recognizer, Object offendingSymbol, int line,
-					int charPositionInLine, String msg, RecognitionException e) {
-				throw new IllegalStateException("failed to parse at line " + line + " due to " + msg, e);
-			}
-		});
-		try {
-			return p.hgvs_variant();
-		} catch (IllegalStateException e) {
-			throw new IllegalStateException("Could not parse \"" + inputString + "\"", e);
-		}
-	}
+public class TestProteinParsing extends ParsingTestBase {
 
 	@Test
 	public void testProteinSingleVarTrivialChanges() {
